split unit tests into multiple functions
